requirements.txt

viam-sdk==0.62.0

typing-extensions
